requirements_dev.txt
-r requirements.txt
pykaa==0.5.2
pykaa-dataproc>=0.2.7,<1
tox==3.25.1
requirements-parser==0.5.0
coverage==4.5.4
pytest-cov==2.10.1
pytest-json-report==1.2.4
pytest-bdd==3.4.0
pytest-cookies==0.6.1
behave~=1.2.6
